fixed missing file error in ZWaveLib project
