+++ MigFiles/SupportLibraries/ZWaveLib/ZWaveEvent.cs
@@ -24,6 +24,54 @@ using System;
 
 namespace ZWaveLib
 {
+    public enum EventParameter
+    {
+        Level,
+        ManufacturerSpecific,
+        Generic,
+        MeterKwHour,
+        MeterKvaHour,
+        MeterWatt,
+        MeterPulses,
+        MeterAcVolt,
+        MeterAcCurrent,
+        MeterPower,
+        SensorTemperature,
+        SensorHumidity,
+        SensorLuminance,
+        SensorMotion,
+        AlarmDoorWindow,
+        AlarmGeneric,
+        AlarmSmoke,
+        AlarmCarbonMonoxide,
+        AlarmCarbonDioxide,
+        AlarmHeat,
+        AlarmFlood,
+        AlarmTampered,
+        Configuration,
+        WakeUpInterval,
+        WakeUpNotify,
+        Association,
+        Battery,
+        NodeInfo,
+        MultiinstanceSwitchBinaryCount,
+        MultiinstanceSwitchBinary,
+        MultiinstanceSwitchMultilevelCount,
+        MultiinstanceSwitchMultilevel,
+        MultiinstanceSensorBinaryCount,
+        MultiinstanceSensorBinary,
+        MultiinstanceSensorMultilevelCount,
+        MultiinstanceSensorMultilevel,
+        ThermostatFanMode,
+        ThermostatFanState,
+        ThermostatHeating,
+        ThermostatMode,
+        ThermostatOperatingState,
+        ThermostatSetBack,
+        ThermostatSetPoint,
+        UserCode
+    }
+
     public class ZWaveEvent
     {
         public ZWaveNode Node;
